# backend/specs/openapi.yaml
openapi: 3.0.3
info:
  title: 智能测试管理平台 API
  description: |
    智能测试管理平台 RESTful API 定义。

    ## 认证
    所有 API（除健康检查和认证接口外）需要 JWT Token 认证：
    ```
    Authorization: Bearer <token>
    ```

    ## 错误响应
    统一错误响应格式：
    ```json
    {
      "error": "错误信息"
    }
    ```
  version: 1.0.0
  contact:
    name: AITestOS Team
    email: support@aitestos.com
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.aitestos.com/api/v1
    description: 生产环境
...
paths:
...
  /generation/drafts:
    get:
      tags: [Generation]
      summary: 获取全局草稿列表
      operationId: listAllDrafts
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
          description: 项目 ID
        - name: taskId
          in: query
          schema:
            type: string
            format: uuid
          description: 任务 ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, rejected]
          description: 草稿状态
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: 偏移量
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: 每页数量
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaseDraft'
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
  /generation/drafts/{id}/confirm:
    post:
      tags: [Generation]
      summary: 确认草稿
      description: 确认草稿并将其转为正式测试用例
      operationId: confirmDraft
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 草稿 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - module_id
              properties:
                module_id:
                  type: string
                  format: uuid
                  description: 关联的模块 ID
      responses:
        '201':
          description: 确认成功，返回创建的测试用例
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCase'
        '409':
          description: 草稿已确认或已拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /generation/drafts/{id}/reject:
    post:
      tags: [Generation]
      summary: 拒绝草稿
      operationId: rejectDraft
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 草稿 ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  enum: [duplicate, irrelevant, low_quality, other]
                  description: 拒绝原因类型
                feedback:
                  type: string
                  description: 拒绝原因反馈
      responses:
        '200':
          description: 拒绝成功
        '409':
          description: 草稿已确认或已拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /knowledge/documents:
    get:
      tags: [Documents]
      summary: 获取文档列表
      operationId: listDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/DocumentType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DocumentStatus'
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
    post:
      tags: [Documents]
      summary: 上传文档
      operationId: uploadDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadDocumentRequest'
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: 无效的文档类型
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /knowledge/documents/{id}:
    get:
      tags: [Documents]
      summary: 获取文档详情
      operationId: getDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
        '404':
          description: 文档不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Documents]
      summary: 删除文档
      operationId: deleteDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CaseStatus:
      type: string
      enum: [unexecuted, pass, block, fail]
    CaseType:
      type: string
      enum: [functionality, performance, api, ui, security]
...
    Priority:
      type: string
      enum: [P0, P1, P2, P3]
...
    DraftStatus:
      type: string
      enum: [pending, confirmed, rejected]
    DocumentType:
      type: string
      enum: [prd, figma, api_spec, swagger, markdown]
    DocumentStatus:
      type: string
      enum: [pending, processing, completed, failed]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 错误信息

    # ==================== 认证 ====================
...
    TestCase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        moduleId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        preconditions:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        expected:
          type: object
        caseType:
          $ref: '#/components/schemas/CaseType'
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/CaseStatus'
        aiMetadata:
          type: object
          description: AI 生成元数据
          properties:
            generationTaskId:
              type: string
              format: uuid
            confidence:
              type: string
              enum: [high, medium, low]
            referencedChunks:
              type: array
              items:
                type: object
                properties:
                  chunkId:
                    type: string
                    format: uuid
                  documentId:
                    type: string
                    format: uuid
                  documentTitle:
                    type: string
                  similarityScore:
                    type: number
                    format: double
            modelVersion:
              type: string
            generatedAt:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CaseDraft:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        title:
          type: string
        preconditions:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        expected:
          type: object
        caseType:
          $ref: '#/components/schemas/CaseType'
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/DraftStatus'
        feedback:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/DocumentType'
        url:
          type: string
        contentText:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/DocumentType'
        url:
          type: string
        contentText:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        chunk_count:
          type: integer
          description: 文档分块数量
    UploadDocumentRequest:
      type: object
      required: [project_id, name, type]
      properties:
        project_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 255
        type:
          type: string
          enum: [prd, figma, api_spec, swagger, markdown]
    DocumentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
...